Fixed Meta tag Error
There was an error with making the layout responsive to screen size changes. The error was found in a lack of the meta tag that describes the viewport. After adding this tag, much of the previous CSS code that used px measurements were no longer accurate and so were resized.